Modify the types of primary key and add some attributes in MenuEntity, OrderEntity and UserAddressEntity. Add enums. Remove the dependency version of 'jaxb-api'.

+++ pizza-data/src/main/java/edu.ecnu.scsse.pizza.data/repository/DeliverIngredientLoggerJpaRepository.java
@@ -3,6 +3,6 @@ package edu.ecnu.scsse.pizza.data.repository;
 import edu.ecnu.scsse.pizza.data.domain.DeliverIngredientLoggerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeliverIngredientLoggerJpaRepository extends JpaRepository<DeliverIngredientLoggerEntity,Long> {
+public interface DeliverIngredientLoggerJpaRepository extends JpaRepository<DeliverIngredientLoggerEntity,Integer> {
 
 }
