+++ pizza-data/src/main/java/edu.ecnu.scsse.pizza.data/repository/IngredientJpaRepository.java
@@ -3,6 +3,6 @@ package edu.ecnu.scsse.pizza.data.repository;
 import edu.ecnu.scsse.pizza.data.domain.IngredientEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IngredientJpaRepository extends JpaRepository<IngredientEntity,Long> {
+public interface IngredientJpaRepository extends JpaRepository<IngredientEntity,Integer> {
 
 }
