+++ pizza-data/src/main/java/edu.ecnu.scsse.pizza.data/repository/OperateLoggerJpaRepository.java
@@ -3,6 +3,6 @@ package edu.ecnu.scsse.pizza.data.repository;
 import edu.ecnu.scsse.pizza.data.domain.OperateLoggerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OperateLoggerJpaRepository extends JpaRepository<OperateLoggerEntity,Long> {
+public interface OperateLoggerJpaRepository extends JpaRepository<OperateLoggerEntity,Integer> {
 
 }
